backend/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
workos==5.21.0
python-dotenv==1.1.0
requests==2.32.3
PyJWT[crypto]==2.10.1
cryptography==45.0.2
mysql-connector-python==9.3.0
psycopg2-binary==2.9.10
pytest
pytest-flask
